refact: 여행 취향 선택 UI 같이 사용할 수 있게 변경 #104 - onboard 페이지에서 사용하던 여행 취향 선택 화면, 일정 생성 페이지에서도 사용할 수 있도록 수정
