+++ src/app/(provider)/(root)/onboard/_components/OnBoard/OnBoard.tsx
@@ -1,36 +1,14 @@
 "use client";
 import SubmitButton from "@/components/commons/SubmitButton";
+import TravelStyle from "@/components/TravelStyle";
 import { useTravelType } from "@/stores/travelType.store";
 import { useRouter } from "next/navigation";
 import { MouseEvent, useEffect, useState } from "react";
-import SelectContainer from "../SelectContainer/SelectContainer";
-import SelectForm from "../SelectForm/SelectForm";
 
 function OnBoard() {
   const { travelType } = useTravelType();
   const router = useRouter();
   const [isValidated, setIsValidated] = useState<boolean>(true);
-
-  const theme = [
-    "힐링",
-    "액티비티",
-    "식도락",
-    "쇼핑",
-    "문화체험",
-    "전시관람",
-    "자연",
-    "도심",
-  ];
-  const season = ["봄", "여름", "가을", "겨울"];
-  const travelMate = [
-    "혼자",
-    "친구와",
-    "연인과",
-    "가족과",
-    "부모님 모시고",
-    "아이와",
-    "반려동물과",
-  ];
 
   useEffect(() => {
     if (
@@ -62,18 +40,9 @@ function OnBoard() {
           <br /> 맞춤설정
         </h1>
         <form className="flex flex-col items-center">
-          <SelectContainer title="여행 테마가 무엇인가요?">
-            <SelectForm buttonList={theme} category="theme"></SelectForm>
-          </SelectContainer>
-          <SelectContainer title="언제 떠나시나요?">
-            <SelectForm buttonList={season} category="season"></SelectForm>
-          </SelectContainer>
-          <SelectContainer title="누구와 떠나시나요?">
-            <SelectForm
-              buttonList={travelMate}
-              category="travelMate"
-            ></SelectForm>
-          </SelectContainer>
+          <TravelStyle title="여행 테마가 무엇인가요?" category="theme" />
+          <TravelStyle title="언제 떠나시나요?" category="season" />
+          <TravelStyle title="누구와 떠나시나요?" category="travelMate" />
           <SubmitButton
             theme="primary"
             size="lg"
