+++ src/components/TravelStyle/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./TravelStyle";
+
